+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Features
 - Support`RevitApplication` static class.
 ### Updated
-- Add `RevitApplication` with static `UIApplication`, `UIControlledApplication` and `IsAddInContext`.
+- Add `RevitApplication` with static `UIApplication`, `UIControlledApplication` and `IsInAddInContext`.
 - Update `StackTraceUtils` to support repack using `IsAssemblyRepack`.
 ### Tests
 - Test `RevitApplication` static class.